Refactor codebase to flatten modules and reduce file count
- Flatten `src/ast/` to `src/ast.rs`
- Flatten `src/codegen/` to `src/codegen.rs`
- Flatten `src/errors/` to `src/errors.rs` (merged `assembly.rs` and `messages.rs`)
- Simplify `src/morphology/`: moved `matcher` logic to `mod.rs`
- Consolidate `src/semantic/`: merged `assembly_model.rs` and tests into `assembler.rs` and `conversion.rs`
- Update `.jules/razor.md` with reduction log
- Ensure all tests pass and clippy is happy
